Combine client_handler_gtk.cpp into client_handler.cpp
No reason to have separate gtk-specific client
handler extension since entire source tree is
Linux specific.
